Started adding 'complex' functionality to the lag correlator.
First, a 'complex inputs' mask parameter was added to all the upper
level blocks and properly propagated to the necessary constituent
blocks.

Then, a 2-bit complex multiplier block was added and a conditional
statement was added to 'single_lag_init.m' that draws the mults using
that block instead of a single 2-bit multiplier. This means that
for now complex inputs will only be supported in the special 2-bit
case.

Next, add a complex add/sub block for the adder tree and a complex
simple accumulator (just two simple_acc's) for the first stage integ-
gration.