Use HDL for the 'overall_latency' delay block
This hopefully allows the cascaded delays to be coalesced into fewer resources.
Even with this, using cascaded delays seems to be somewhat inefficient compared
to using a single delay of the desired total length.  Another improvement would
be to put the single delay of the desired total length just before the inputs
to each lag's multipliers where it would be narrower than after the
product-sum.